feat: 모든 기물들의 위치를 저장하는 보드 클래스 추가
