+++ README.md
@@ -13,8 +13,8 @@
   - [x] 모든 좌표 값을 가진다. (초기화)
   - [x] 기물을 초기 위치에 생성한다
 
-- [ ] 보드
-  - [ ] 모든 기물들의 위치를 저장한다.
+- [x] 보드
+  - [x] 모든 기물들의 위치를 저장한다.
 
   
 - [x] 팀
