+++ src/test/java/chessgame/BoardTest.java
@@ -1,4 +1,4 @@
-package chessgame.board;
+package chessgame;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -8,6 +8,6 @@ class BoardTest {
     @Test
     @DisplayName("모든 좌표값을 가진 보드를 생성한다")
     void Should_NotThrowException_When_ConstructBord() {
-        Assertions.assertDoesNotThrow(Board::create);
+        Assertions.assertDoesNotThrow(() -> new Board(ChessBoardFactory.create()));
     }
 }
